% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_Functions.R
\name{app_about}
\alias{app_about}
\title{app_about}
\usage{
app_about(about.link, about.who, about.text)
}
\arguments{
\item{about.link}{A link to their page, typically finance or city manager or maybe they have a PBB page}

\item{about.who}{About who...}

\item{about.text}{html of some paragraphs on why PBB}
}
\description{
Does the layout of the about section
}
\examples{
app_about(about.link,about.who,about.text)
}
