% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/UI_Functions.R
\name{app_charts}
\alias{app_charts}
\title{app_charts}
\usage{
app_charts(hasMetrics = F)
}
\arguments{
\item{hasMetrics}{Include the metrics tab}
}
\description{
Sets the UI of charts for our app across a set of tabs
}
\examples{
app_charts()
}

