% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_Functions.R
\name{app_footer}
\alias{app_footer}
\title{app_footer}
\usage{
app_footer(logo, logo.top_margin = 20, height = 205,
  background_color = "#eee", rx_logo = "resourcex_logo.png",
  rx_logo.top_margin = 35, copyright = "2018 Copyright City of")
}
\arguments{
\item{logo}{org footer logo}

\item{logo.top_margin}{some spacing for the org logo to top of the footer}

\item{height}{footer height}

\item{background_color}{color of the bottom footer}

\item{rx_logo}{resourcex logo saved in assets}

\item{rx_logo.top_margin}{some spacing between org logo and the rx logo at the bottom}

\item{copyright}{copyright text}
}
\description{
Does the layout of the tabs and loads organization header image. Also set some locations of our info modal and optional org logo
}
\examples{
app_header(header='header.jpg',info.top=245,info.left=20,header_logo=NULL,header_logo.top=NULL,header_logo.left=NULL,tabs.height=225)
}
