% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_Functions.R
\name{app_head}
\alias{app_head}
\title{app_head}
\usage{
app_head(title = "PBB", favicon = "favicon.ico",
  base = "https://openpbbdata.net/present")
}
\arguments{
\item{title}{pAGE title}

\item{db_host_new}{database host}
}
\description{
loads libraries and sets page title and favicon
}
\examples{
page_head(title="PBB",favicon="favicon.ico")
}
