% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_Functions.R
\name{app_header}
\alias{app_header}
\title{app_header}
\usage{
app_header(header = "header.jpg", info.top = 245, info.left = 20,
  header_logo = NULL, header_logo.top = NULL,
  header_logo.left = NULL, tabs.height = 225, info.click_me = NULL,
  info.intro = NULL)
}
\arguments{
\item{header}{page header image}

\item{info.top}{position of info button from the top}

\item{info.left}{position of info button from the left}

\item{header_logo}{optional org logo for header}

\item{header_logo.top}{optional org logo position from top}

\item{header_logo.left}{optional org logo position from left}

\item{tabs.height}{height of tab widget}

\item{info.click_me}{additional text can be HTML formatted with style next to info icon}

\item{info.intro}{run the intro tutorial on page load. False or Null requires user to click info icon in header.}
}
\description{
Does the layout of the tabs and loads organization header image. Also set some locations of our info modal and optional org logo
}
\examples{
app_header(header='header.jpg',info.top=245,info.left=20,header_logo=NULL,header_logo.top=NULL,header_logo.left=NULL,tabs.height=225)
}
