% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeSummaries.R
\name{colnameIDs_to_Names}
\alias{colnameIDs_to_Names}
\title{colnameIDs_to_Names}
\usage{
colnameIDs_to_Names(df.update, df.LinkID_Name, ID, Name)
}
\arguments{
\item{df.update}{dataframe to update}
}
\description{
Helper Function
}
