% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeSummaries.R
\name{create_IDstring}
\alias{create_IDstring}
\title{create_IDstring}
\usage{
create_IDstring(numeric_vector)
}
\arguments{
\item{numeric_vector}{vector to format to convert into MySQL WHERE IN format. Can also be character vector}
}
\description{
Helper Function
}
