% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadBudgetData.R
\name{getSummaryAll}
\alias{getSummaryAll}
\title{getSummaryAll}
\usage{
getSummaryAll(DropBox_dir, local_dir, budget = input$budget_year,
  tab = input$tabset, use_DropBox = T)
}
\arguments{
\item{DropBox_dir}{the directory on dropbox for the organization}

\item{budget}{what budget to use, the csv file should be in a folder of the same name}

\item{tab}{what tabset to use from our radio buttons, the config folder should have the tab files in json format ending in "_tabs.json"}

\item{session}{shiny server session variable}

\item{values}{reactiveValue that stores the summary data}
}
\description{
sends the tab data and treemap data to the client as well as updates the summary all data for program summary
}
\examples{
getSummaryAll(DropBox_dir,local_dir,budget=input$budget_year,tab=input$tabset,use_DropBox=T)
}
