% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DropBoxUpdate.R
\name{getTabData}
\alias{getTabData}
\title{getTabData}
\usage{
getTabData(DropBox_dir, local_dir, tab = input$tabset,
  budget = input$budget_year, use_DropBox = T)
}
\arguments{
\item{DropBox_dir}{the directory on dropbox for the organization}

\item{tab}{what tabset to use from our radio buttons, the config folder should have the tab files in json format ending in "_tabs.json"}

\item{budget}{what budget to use, the csv file should be in a folder of the same name}
}
\description{
reads the json file for the tabset of interest
}
\examples{
getTabData(DropBox_dir,tab=tab,budget=budget)
}

