% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DropBoxUpdate.R
\name{getTreeMapData_CSV}
\alias{getTreeMapData_CSV}
\title{getTreeMapData_CSV}
\usage{
getTreeMapData_CSV(DropBox_dir, local_dir, budget = input$budget,
  tabset = input$tabset, data_treemap = "/data_treemap.csv",
  use_DropBox = T)
}
\arguments{
\item{DropBox_dir}{the directory on dropbox for the organization}

\item{budget}{what budget to use, the csv file should be in a folder of the same name}

\item{tabset}{path to results tabs json file}

\item{data_treemap}{filename to use for the csv data}
}
\description{
reads the csv file for the treemap from dropbox
}
\examples{
getTreeMapData_CSV(DropBox_dir,budget=budget,data_treemap="/data_treemap.csv")
}
