% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeProgramResultsTable.R
\name{make_ProgramResultsTable}
\alias{make_ProgramResultsTable}
\title{make_ProgramResultsTable}
\usage{
make_ProgramResultsTable(df, buttons = c("copy", "excel", "print"))
}
\arguments{
\item{df}{present pbb data}

\item{buttons}{what buttons to use}
}
\description{
Create a DT table of the pbb data for the chart tabs
}
\examples{
make_ProgramResultsTable(df,buttons=c('copy', 'excel', 'print'))
}
