% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadBudgetData.R
\name{makeTreeMapData}
\alias{makeTreeMapData}
\title{makeTreeMapData}
\usage{
makeTreeMapData(df, Tabdata, TabIndex = 0)
}
\arguments{
\item{df}{the csv file to convert to json for the tree map}

\item{Tabdata}{the json tab data we send along with tree map data to update the tab plus tree map widget}
}
\description{
reads the csv file for the treemap from dropbox and converts to JSON
}
\examples{
getTreeMapData(DropBox_dir,budget=budget,data_treemap="/data_treemap.csv")
}
