% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_Functions.R
\name{page_head}
\alias{page_head}
\title{page_head}
\usage{
page_head(title = "PBB", favicon = "favicon.ico")
}
\arguments{
\item{title}{pAGE title}

\item{db_host_new}{database host}
}
\description{
loads libraries and sets page title and favicon
}
\examples{
page_head(title="PBB",favicon="favicon.ico")
}
