% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_Functions.R
\name{page_spinner}
\alias{page_spinner}
\title{page_spinner}
\usage{
page_spinner(spinner = "spin4.gif")
}
\arguments{
\item{spinner}{spinner.gif file saved in assets}
}
\description{
put this in the head funtion to use a busy spinner. USees gif file saved in assets
}
\examples{
page_spinner(spinner='spin4.gif')


}
