% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadBudgetData.R
\name{present_initialize}
\alias{present_initialize}
\title{present_initialize}
\usage{
present_initialize(session, use_DropBox = FALSE, local_dir = "./budgets/",
  DropBox_dir = NULL)
}
\arguments{
\item{session}{shiny server session variable}

\item{use_DropBox}{essentially depricated. we moved to local file system.. Exploring moving to AWS S3 buckets}

\item{local_dir}{the local directory where budgets are kept}

\item{DropBox_dir}{the directory on dropbox for the organization}
}
\description{
initializes the dropdowns on the page and gets the summary data files
}
\examples{
present_initialize(session,use_DropBox=FALSE,local_dir="./budgets/",DropBox_dir=NULL)
}
