% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProgramModalSummary.R
\name{ProgramModal}
\alias{ProgramModal}
\title{Program Summary Modal}
\usage{
ProgramModal(Modal_header = T, Modal_tabs = T, TotalCost_tab = T,
  Positions_tab = T, OperatingCosts_tab = T,
  Program = input$chartdata_Program, Desc = input$chartdata_Desc,
  TotalCost = input$chartdata_TotalCost, FTE = input$chartdata_FTE)
}
\arguments{
\item{Modal_header}{Summary of data echo'd back from the csv setup file}

\item{TotalCost}{Pie chart of Personnel and NonPersonnel}

\item{Positions}{Include Position Pie and Table}

\item{OperatingCosts}{Include operating cost pie and table}
}
\description{
Creates a modal summary of a program. You may select to only include a header summary or select what tabs to include.
}
\examples{
showModal(ProgramModal(Modal_header=T,Modal_tabs=F,TotalCost=T,Positions=T,OperatingCosts=T))
showModal(ProgramModal(Modal_header=T,Modal_tabs=T,TotalCost=T,Positions=T,OperatingCosts=T))
}
