% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeSummaries.R
\name{RadSlicer_OpenData}
\alias{RadSlicer_OpenData}
\title{RadSlicer_OpenData}
\usage{
RadSlicer_OpenData(con, BudgetID, CostModelID)
}
\arguments{
\item{con}{RMySQL connection}

\item{BudgetID}{BudgetID to use}

\item{CostModelID}{PBB CostModel ID}
}
\description{
Helper Function
}
