% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ProgramModalSummary.R
\name{rhandson_ProgramCostSummary}
\alias{rhandson_ProgramCostSummary}
\title{rhandson_ProgramCostSummary}
\usage{
rhandson_ProgramCostSummary(df, height = 175)
}
\arguments{
\item{df}{Cost data to summarize}

\item{height}{Table height}
}
\description{
Makes a cost summary tabel for our Program Summary Modal
}
\examples{
showModal(ProgramModal(Modal_header=T,Modal_tabs=F,TotalCost=T,Positions=T,OperatingCosts=T))
}

