% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gVis_Pie.R
\name{rollup_gvisPie}
\alias{rollup_gvisPie}
\title{helper function for making data for gvis plots}
\usage{
rollup_gvisPie(df, labelvar, numvar)
}
\arguments{
\item{df}{Data to roll up}

\item{labelvar}{label variable, example "Personnel"}

\item{numvar}{Number to sum, example "Cost"}
}
\description{
Sends email using mailR
}
\examples{
rollup_gvisPie(df,labelvar='Cost Type',numvar='Cost')
}
