% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize.R
\name{update_ReportsTab}
\alias{update_ReportsTab}
\title{update_ReportsTab}
\usage{
update_ReportsTab(local_dir, budget)
}
\arguments{
\item{local_dir}{the local directory where budgets are kept}

\item{budget}{what budget to use, the csv file should be in a folder of the same name}
}
\description{
Checks for any reports and adds or removes tab to main tabset of charts
}
\examples{
update_ReportsTab(local_dir,budget)
}
