% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateFunctions.R
\name{update_Tabs}
\alias{update_Tabs}
\title{update_Tabs}
\usage{
update_Tabs(tabset, bpas, results)
}
\arguments{
\item{tabset}{input selector}

\item{bpas}{path to bpas tabs json file}

\item{results}{path to results tabs json file}
}
\description{
updates the tabset
}
\examples{
update_Tabs(tabset=input$tabset,bpas='./www/config/roswell_bpas.json',results='./www/config/roswell_tabs.json')
}
