% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_Functions.R
\name{app_charts}
\alias{app_charts}
\title{app_charts}
\usage{
app_charts(results_tab = NULL, results_chart_height = 500,
  hasMetrics = NULL)
}
\arguments{
\item{results_tab}{Include the summary results tab}

\item{results_chart_height}{height of summary chart}

\item{hasMetrics}{Include the metrics tab}
}
\description{
Sets the UI of charts for our app across a set of tabs
}
\examples{
app_charts()
}
