% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DropBoxUpdate.R
\name{getTreeMapData}
\alias{getTreeMapData}
\title{getTreeMapData}
\usage{
getTreeMapData(DropBox_dir, budget = input$budget, tabset = input$tabset,
  data_treemap = "/data_treemap.csv")
}
\arguments{
\item{DropBox_dir}{the directory on dropbox for the organization}

\item{budget}{what budget to use, the csv file should be in a folder of the same name}

\item{tabset}{path to results tabs json file}

\item{data_treemap}{filename to use for the csv data}
}
\description{
reads the csv file for the treemap from dropbox
}
\examples{
getTreeMapData(DropBox_dir,budget=budget,data_treemap="/data_treemap.csv")
}
