% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProgramModalSummary.R
\name{gVis_Pie_ProgramCostSummary}
\alias{gVis_Pie_ProgramCostSummary}
\title{Function for making gVis Cost summary of Total Cost, Personnel, or NonPersonnel}
\usage{
gVis_Pie_ProgramCostSummary(df, dataTitle = "Total Cost: ",
  labelvar = "Cost Type", numvar = "Cost",
  sliceVisibilityThreshold = NULL)
}
\arguments{
\item{df}{Program summary data}

\item{dataTitle}{Title for cost summary. Example "Total Cost: " or "Personnel: "}

\item{labelvar}{label variable, example "Personnel"}

\item{numvar}{Number to sum, example "Cost"}
}
\description{
makes a plot
}
\examples{
rollup_gvisPie(df,labelvar='Cost Type',numvar='Cost')
}
