% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeSummaries.R
\name{PullAllocations_OpenData}
\alias{PullAllocations_OpenData}
\title{PullAllocations_OpenData}
\usage{
PullAllocations_OpenData(conexists = con, BudgetID, CostModelID)
}
\arguments{
\item{con}{RMySQL connection}
}
\description{
Helper Function
}
