% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeSummaries.R
\name{PullFinalScores_OpenData}
\alias{PullFinalScores_OpenData}
\title{PullAllocations_OpenData}
\usage{
PullFinalScores_OpenData(conexists = NULL, BudgetID, CostModelID)
}
\arguments{
\item{conexists}{RMySQL connection}
}
\description{
Helper Function
}
