% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitFeedback.R
\name{submitFeedback}
\alias{submitFeedback}
\title{A function to send feedback}
\usage{
submitFeedback(name = input$name, email = input$email,
  comment = input$comment, emails = emails)
}
\arguments{
\item{name}{Sender's name}

\item{email}{Sender's email}

\item{comment}{Email's content, the feedback they submitted}

\item{emails}{vector of emails to send feedback to}
}
\description{
Snds email using mailR
}
\examples{
submitFeedback()

}
