% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/submitFeedback.R
\name{submitFeedback}
\alias{submitFeedback}
\title{A function to send feedback}
\usage{
submitFeedback(name = input$name, email = input$email,
  comment = input$comment, emails = emails, Subject)
}
\arguments{
\item{name}{Sender's name}

\item{email}{Sender's email}

\item{comment}{Email's content, the feedback they submitted}

\item{emails}{vector of emails to send feedback to}
}
\description{
Sends email using mailR
}
\examples{
submitFeedback()
}

