% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProgramModalSummary.R
\name{subset_SummaryAll}
\alias{subset_SummaryAll}
\title{subset_SummaryAll}
\usage{
subset_SummaryAll(df, select_by, filter, Cost.Type = NULL)
}
\arguments{
\item{df}{Data to subset}

\item{select_by}{column to select on.}

\item{filter}{row element to choose. Example, Prg# = 510}

\item{Cost.Type}{Option to subset by Personnel or NonPersonnel}
}
\description{
subsets data
}
\examples{
subset_SummaryAll(df=SummaryAll,select_by='Prg#',filter=chartdata_Program(),Cost.Type=NULL)
}
