% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DropBoxUpdate.R
\name{updateDropBoxBudget}
\alias{updateDropBoxBudget}
\title{updateDropBoxBudget}
\usage{
updateDropBoxBudget(budget = input$budget_year, tab = input$tabset,
  DropBox_dir)
}
\arguments{
\item{budget}{what budget to use, the csv file should be in a folder of the same name}

\item{tab}{what tabset to use from our radio buttons, the config folder should have the tab files in json format ending in "_tabs.json"}

\item{DropBox_dir}{the directory on dropbox for the organization}
}
\description{
sends the tab data and treemap data to the client as well as updates the summary all data for program summary
}
\examples{
updateDropBoxBudget(budget=input$budget_year,tab=input$tabset,DropBox_dir)
}
